+++ plugins/design-system-wordpress-plugin/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 -  Added new banner functionality ([DSWP-2](https://apps.itsm.gov.bc.ca/jira/browse/DSWP-2))
 -  Added content security policy functionality ([DSWP-49](https://apps.itsm.gov.bc.ca/jira/browse/DSWP-49))
+-  Added skip navigation functionality ([DSWP-51](https://apps.itsm.gov.bc.ca/jira/browse/DSWP-51))